Make sure you recreate only CachesInDB

+++ objects/mysql_dal.php
@@ -204,7 +204,7 @@ class sqlDAL
              */
 
              _error_log("writeSql [{$stmt->errno}] {$stmt->error} ".' '.json_encode(debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS)));
-             if($stmt->errno == 1205){//Lock wait timeout exceeded; try restarting transaction
+             if($stmt->errno == 1205 && preg_match('/CachesInDB/', $preparedStatement)){//Lock wait timeout exceeded; try restarting transaction
                 _error_log("writeSql Recreate CachesInDB ");
                 $sql = 'DROP TABLE IF EXISTS `CachesInDB`';
                 $global['mysqli']->query($sql);
